// game/types/core.d.ts
export type Vec2 = [number, number]
export type Vec3 = [number, number, number]
export type SpriteFrame = [x: number, y: number, width: number, height: number]
export type ColorTriplet = [string, string, string]

export type ResourceAmounts = [
	number,
	number,
	number,
	number,
	number,
	number,
	number,
	number,
	number,
	number,
]